+++ src/main/java/it/eng/idsa/businesslogic/service/AuditEventService.java
@@ -4,12 +4,14 @@ import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.time.LocalTime;
 import java.util.List;
+import java.util.stream.Collectors;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import it.eng.idsa.businesslogic.entity.AuditLog;
 import it.eng.idsa.businesslogic.repository.AuditEventRepository;
+import it.eng.idsa.businesslogic.util.AES256Static;
 
 @Service
 public class AuditEventService {
@@ -21,13 +23,27 @@ public class AuditEventService {
 	}
 
 	public List<AuditLog> getAllAuditEvents() {
-		return auditRepository.findAll();
+		return auditRepository.findAll()
+				.parallelStream()
+				.map(this::decryptAuditLog)
+				.collect(Collectors.toList());
 	}
 
 	public List<AuditLog> getAuditEventsForDate(LocalDate date) {
 		LocalDateTime startOfDay = date.atStartOfDay(); // Start of the day
 		LocalDateTime endOfDay = date.atTime(LocalTime.MAX); // End of the day
 
-		return auditRepository.findByTimestampBetween(startOfDay, endOfDay);
+		return auditRepository.findByTimestampBetween(startOfDay, endOfDay)
+				.parallelStream()
+				.map(this::decryptAuditLog)
+				.collect(Collectors.toList());
+	}
+	
+	private AuditLog decryptAuditLog(AuditLog auditLog) {
+		AuditLog a = new AuditLog();
+		a.setId(auditLog.getId());
+		a.setEvent(AES256Static.decrypt(auditLog.getEvent()));
+		a.setTimestamp(auditLog.getTimestamp());
+		return a;
 	}
 }